Fix signatures.

+++ src/lte/model/epc-ue-nas.cc
@@ -100,14 +100,14 @@ EpcUeNas::GetTypeId (void)
   return tid;
 }
 
-void 
+void
 EpcUeNas::SetDevice (Ptr<NetDevice> dev)
 {
   NS_LOG_FUNCTION (this << dev);
   m_device = dev;
 }
 
-void 
+void
 EpcUeNas::SetImsi (uint64_t imsi)
 {
   NS_LOG_FUNCTION (this << imsi);
@@ -158,13 +158,13 @@ EpcUeNas::SetForwardUpCallback (Callback <void, Ptr<Packet> > cb)
 }
 
 void
-EpcUeNas::StartCellSelection (uint16_t dlEarfcn)
+EpcUeNas::StartCellSelection (uint32_t dlEarfcn)
 {
   NS_LOG_FUNCTION (this << dlEarfcn);
   m_asSapProvider->StartCellSelection (dlEarfcn);
 }
 
-void 
+void
 EpcUeNas::Connect ()
 {
   NS_LOG_FUNCTION (this);
@@ -174,7 +174,7 @@ EpcUeNas::Connect ()
 }
 
 void
-EpcUeNas::Connect (uint16_t cellId, uint16_t dlEarfcn)
+EpcUeNas::Connect (uint16_t cellId, uint32_t dlEarfcn)
 {
   NS_LOG_FUNCTION (this << cellId << dlEarfcn);
 
@@ -201,7 +201,7 @@ EpcUeNas::ConnectMc (uint16_t cellId, uint16_t dlEarfcn, uint16_t mmWaveCellId)
 }
 
 
-void 
+void
 EpcUeNas::Disconnect ()
 {
   NS_LOG_FUNCTION (this);
@@ -210,7 +210,7 @@ EpcUeNas::Disconnect ()
 }
 
 
-void 
+void
 EpcUeNas::ActivateEpsBearer (EpsBearer bearer, Ptr<EpcTft> tft)
 {
   NS_LOG_FUNCTION (this);
@@ -247,7 +247,7 @@ EpcUeNas::Send (Ptr<Packet> packet)
           }
         else
           {
-            m_asSapProvider->SendData (packet, bid); 
+            m_asSapProvider->SendData (packet, bid);
             return true;
           }
       }
@@ -260,13 +260,13 @@ EpcUeNas::Send (Ptr<Packet> packet)
     }
 }
 
-void 
+void
 EpcUeNas::DoNotifyConnectionSuccessful (uint16_t rnti)
 {
   NS_LOG_FUNCTION (this);
   switch (m_state)
   {
-    case ACTIVE: // this means the Master LTE Cell was already connected 
+    case ACTIVE: // this means the Master LTE Cell was already connected
       {
         // notify the LTE eNB RRC that a secondary cell is available
         m_asSapProvider->NotifySecondaryCellConnected(rnti, m_mmWaveCellId);
@@ -286,7 +286,7 @@ EpcUeNas::DoNotifyHandoverSuccessful (uint16_t rnti, uint16_t mmWaveCellId)
   NS_LOG_FUNCTION (this);
   switch (m_state)
   {
-    case ACTIVE: // this means the Master LTE Cell was already connected 
+    case ACTIVE: // this means the Master LTE Cell was already connected
       {
         // notify the LTE eNB RRC that a secondary cell is available
         m_asSapProvider->NotifySecondaryCellConnected(rnti, m_mmWaveCellId);
@@ -296,13 +296,13 @@ EpcUeNas::DoNotifyHandoverSuccessful (uint16_t rnti, uint16_t mmWaveCellId)
     default:
       SwitchToState (ACTIVE); // will eventually activate dedicated bearers
       break;
-  } 
+  }
 }
 
 void
 EpcUeNas::DoNotifyConnectToMmWave(uint16_t mmWaveCellId)
 {
-  NS_LOG_LOGIC(mmWaveCellId);  
+  NS_LOG_LOGIC(mmWaveCellId);
   m_mmWaveCellId = mmWaveCellId;
 
   if(m_mmWaveAsSapProvider != 0) {
@@ -318,21 +318,21 @@ EpcUeNas::DoNotifyConnectToMmWave(uint16_t mmWaveCellId)
   } else {
     NS_LOG_WARN("Trying to connect to a secondary cell a non MC capable device");
   }
-  
+
 }
 
 void
 EpcUeNas::DoNotifySecondaryCellHandoverStarted (uint16_t oldRnti, uint16_t newRnti, uint16_t mmWaveCellId, LteRrcSap::RadioResourceConfigDedicated rrcd)
 {
   m_mmWaveCellId = mmWaveCellId;
-  NS_ASSERT(m_asSapProvider != 0); 
-  
+  NS_ASSERT(m_asSapProvider != 0);
+
   NS_ASSERT_MSG(mmWaveCellId > 0, "Invalid CellId");
 
   NS_LOG_INFO("Notify the LTE RRC of the secondary cell HO to " << mmWaveCellId);
   // Notify the LTE RRC of the secondary cell HO
   m_asSapProvider->NotifySecondaryCellHandover(oldRnti, newRnti, mmWaveCellId, rrcd);
-   
+
 }
 
 void
@@ -351,14 +351,14 @@ EpcUeNas::DoRecvData (Ptr<Packet> packet)
   m_forwardUpCallback (packet);
 }
 
-void 
+void
 EpcUeNas::DoNotifyConnectionReleased ()
 {
   NS_LOG_FUNCTION (this);
   SwitchToState (OFF);
 }
 
-void 
+void
 EpcUeNas::DoActivateEpsBearer (EpsBearer bearer, Ptr<EpcTft> tft)
 {
   NS_LOG_FUNCTION (this);
@@ -374,7 +374,7 @@ EpcUeNas::GetState () const
   return m_state;
 }
 
-void 
+void
 EpcUeNas::SwitchToState (State newState)
 {
   NS_LOG_FUNCTION (this << ToString (newState));
@@ -403,4 +403,3 @@ EpcUeNas::SwitchToState (State newState)
 
 
 } // namespace ns3
-
