+++ src/lte/model/lte-as-sap.h
@@ -41,7 +41,7 @@ class LteEnbNetDevice;
  * In particular, this class implements the
  * Provider part of the SAP, i.e., the methods exported by the
  * LteUeRrc and called by the EpcUeNas.
- * 
+ *
  */
 class LteAsSapProvider
 {
@@ -61,15 +61,15 @@ public:
    *
    * \param dlEarfcn the downlink carrier frequency (EARFCN)
    */
-  virtual void StartCellSelection (uint16_t dlEarfcn) = 0;
+  virtual void StartCellSelection (uint32_t dlEarfcn) = 0;
 
-  /** 
+  /**
    * \brief Force the RRC entity to stay camped on a certain eNodeB.
    *
    * \param cellId the cell ID identifying the eNodeB
    * \param dlEarfcn the downlink carrier frequency (EARFCN)
    */
-  virtual void ForceCampedOnEnb (uint16_t cellId, uint16_t dlEarfcn) = 0;
+  virtual void ForceCampedOnEnb (uint16_t cellId, uint32_t dlEarfcn) = 0;
 
   /**
    * \brief Tell the RRC entity to enter Connected mode.
@@ -81,7 +81,7 @@ public:
    */
   virtual void Connect (void) = 0;
 
-  /** 
+  /**
    * \brief Send a data packet.
    *
    * \param packet the packet
@@ -90,19 +90,19 @@ public:
   virtual void SendData (Ptr<Packet> packet, uint8_t bid) = 0;
 
 
-  /** 
+  /**
    * \brief Tell the RRC entity to release the connection.
    *
    */
   virtual void Disconnect () = 0;
 
-  /** 
+  /**
    * \brief Tell the RRC that a secondary cell was connected
    *
    */
   virtual void NotifySecondaryCellConnected (uint16_t rnti, uint16_t mmWaveCellId) = 0;
 
-  /** 
+  /**
    * \brief Tell the LTE RRC in the UE that a secondary cell handover was performed,
    * and trigger the RLC instances update
    *
@@ -115,19 +115,19 @@ public:
 /**
  * This class implements the Access Stratum (AS) Service Access Point
  * (SAP), i.e., the interface between the EpcUeNas and the LteUeRrc
- * In particular, this class implements the 
+ * In particular, this class implements the
  * User part of the SAP, i.e., the methods exported by the
  * EpcUeNas and called by the LteUeRrc.
- * 
+ *
  */
 class LteAsSapUser
 {
 public:
   virtual ~LteAsSapUser ();
 
-  /** 
+  /**
    * \brief Notify the NAS that RRC Connection Establishment was successful.
-   * 
+   *
    */
   virtual void NotifyConnectionSuccessful (uint16_t rnti) = 0;
 
@@ -135,28 +135,28 @@ public:
 
   virtual void NotifySecondaryCellHandoverStarted (uint16_t oldRnti, uint16_t newRnti, uint16_t mmWaveCellId, LteRrcSap::RadioResourceConfigDedicated radioResourceConfigDedicated) = 0;
 
-  /** 
+  /**
    * \brief Notify the NAS that LTE RRC received an indication to connect to a MmWave eNB
-   * 
+   *
    */
   virtual void NotifyConnectToMmWave (uint16_t mmWaveCellId) = 0;
 
-  /** 
+  /**
    * \brief Notify the NAS that RRC Connection Establishment failed.
-   * 
+   *
    */
   virtual void NotifyConnectionFailed () = 0;
 
 
-  /** 
+  /**
    * Notify the NAS that RRC Connection was released
-   * 
+   *
    */
   virtual void NotifyConnectionReleased () = 0;
 
-  /** 
+  /**
    * receive a data packet
-   * 
+   *
    * \param packet the packet
    */
   virtual void RecvData (Ptr<Packet> packet) = 0;
@@ -169,7 +169,7 @@ public:
 /**
  * Template for the implementation of the LteAsSapProvider as a member
  * of an owner class of type C to which all methods are forwarded
- * 
+ *
  */
 template <class C>
 class MemberLteAsSapProvider : public LteAsSapProvider
@@ -179,8 +179,8 @@ public:
 
   // inherited from LteAsSapProvider
   virtual void SetCsgWhiteList (uint32_t csgId);
-  virtual void StartCellSelection (uint16_t dlEarfcn);
-  virtual void ForceCampedOnEnb (uint16_t cellId, uint16_t dlEarfcn);
+  virtual void StartCellSelection (uint32_t dlEarfcn);
+  virtual void ForceCampedOnEnb (uint16_t cellId, uint32_t dlEarfcn);
   virtual void Connect (void);
   virtual void SendData (Ptr<Packet> packet, uint8_t bid);
   virtual void Disconnect ();
@@ -212,20 +212,20 @@ MemberLteAsSapProvider<C>::SetCsgWhiteList (uint32_t csgId)
 
 template <class C>
 void
-MemberLteAsSapProvider<C>::StartCellSelection (uint16_t dlEarfcn)
+MemberLteAsSapProvider<C>::StartCellSelection (uint32_t dlEarfcn)
 {
   m_owner->DoStartCellSelection (dlEarfcn);
 }
 
 template <class C>
 void
-MemberLteAsSapProvider<C>::ForceCampedOnEnb (uint16_t cellId, uint16_t dlEarfcn)
+MemberLteAsSapProvider<C>::ForceCampedOnEnb (uint16_t cellId, uint32_t dlEarfcn)
 {
   m_owner->DoForceCampedOnEnb (cellId, dlEarfcn);
 }
 
 template <class C>
-void 
+void
 MemberLteAsSapProvider<C>::Connect ()
 {
   m_owner->DoConnect ();
@@ -239,21 +239,21 @@ MemberLteAsSapProvider<C>::SendData (Ptr<Packet> packet, uint8_t bid)
 }
 
 template <class C>
-void 
+void
 MemberLteAsSapProvider<C>::Disconnect ()
 {
   m_owner->DoDisconnect ();
 }
 
 template <class C>
-void 
+void
 MemberLteAsSapProvider<C>::NotifySecondaryCellConnected (uint16_t rnti, uint16_t mmWaveCellId)
 {
   m_owner->DoNotifySecondaryCellConnected (rnti, mmWaveCellId);
 }
 
 template <class C>
-void 
+void
 MemberLteAsSapProvider<C>::NotifySecondaryCellHandover (uint16_t oldRnti, uint16_t newRnti, uint16_t mmWaveCellId, LteRrcSap::RadioResourceConfigDedicated radioResourceConfigDedicated)
 {
   m_owner->DoNotifySecondaryCellHandover (oldRnti, newRnti, mmWaveCellId, radioResourceConfigDedicated);
@@ -263,7 +263,7 @@ MemberLteAsSapProvider<C>::NotifySecondaryCellHandover (uint16_t oldRnti, uint16
 /**
  * Template for the implementation of the LteAsSapUser as a member
  * of an owner class of type C to which all methods are forwarded
- * 
+ *
  */
 template <class C>
 class MemberLteAsSapUser : public LteAsSapUser
@@ -297,42 +297,42 @@ MemberLteAsSapUser<C>::MemberLteAsSapUser ()
 }
 
 template <class C>
-void 
+void
 MemberLteAsSapUser<C>::NotifyConnectionSuccessful (uint16_t rnti)
 {
   m_owner->DoNotifyConnectionSuccessful (rnti);
 }
 
 template <class C>
-void 
+void
 MemberLteAsSapUser<C>::NotifyHandoverSuccessful (uint16_t rnti, uint16_t mmWaveCellId)
 {
   m_owner->DoNotifyHandoverSuccessful (rnti, mmWaveCellId);
 }
 
 template <class C>
-void 
+void
 MemberLteAsSapUser<C>::NotifyConnectToMmWave (uint16_t mmWaveCellId)
 {
   m_owner->DoNotifyConnectToMmWave (mmWaveCellId);
 }
 
 template <class C>
-void 
+void
 MemberLteAsSapUser<C>::NotifyConnectionFailed ()
 {
   m_owner->DoNotifyConnectionFailed ();
 }
 
 template <class C>
-void 
+void
 MemberLteAsSapUser<C>::RecvData (Ptr<Packet> packet)
 {
   m_owner->DoRecvData (packet);
 }
 
 template <class C>
-void 
+void
 MemberLteAsSapUser<C>::NotifyConnectionReleased ()
 {
   m_owner->DoNotifyConnectionReleased ();
@@ -340,7 +340,7 @@ MemberLteAsSapUser<C>::NotifyConnectionReleased ()
 
 
 template <class C>
-void 
+void
 MemberLteAsSapUser<C>::NotifySecondaryCellHandoverStarted (uint16_t oldRnti, uint16_t newRnti, uint16_t mmWaveCellId, LteRrcSap::RadioResourceConfigDedicated radioResourceConfigDedicated)
 {
   m_owner->DoNotifySecondaryCellHandoverStarted (oldRnti, newRnti, mmWaveCellId, radioResourceConfigDedicated);
